requirements.txt
google-auth
google-cloud-container
kubernetes

